Optimize the code, modify the parse json path
