+++ coinlib/src/main/java/com/keystone/coinlib/v8/ScriptLoader.java
@@ -17,12 +17,8 @@
 
 package com.keystone.coinlib.v8;
 
-import android.content.Context;
 import android.content.res.AssetManager;
-import android.os.storage.StorageManager;
-import android.os.storage.StorageVolume;
 import android.text.TextUtils;
-import android.util.Log;
 
 import com.eclipsesource.v8.V8;
 import com.keystone.coinlib.Coinlib;
@@ -31,13 +27,9 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.lang.reflect.Method;
-import java.util.List;
 
 public class ScriptLoader {
     private static final String TAG = "ScriptLoader";
@@ -92,57 +84,6 @@ public class ScriptLoader {
             e.printStackTrace();
         }
         return stringBuilder.toString();
-    }
-
-    public static String getContentFromSdCard(String path, String fileName) {
-        if (TextUtils.isEmpty(externalSDCardPath())) {
-            Log.i(TAG, "sdCard is not exists");
-            return "";
-        }
-        File file = new File(externalSDCardPath() + File.separator + path, fileName + ".json");
-        if (!file.exists()) {
-            return "";
-        }
-        StringBuilder stringBuilder = new StringBuilder();
-        BufferedReader bfr = null;
-        try {
-            bfr = new BufferedReader(new FileReader(file));
-            String line = bfr.readLine();
-            while (line != null) {
-                stringBuilder.append(line);
-                stringBuilder.append("\n");
-                line = bfr.readLine();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            if (bfr != null) {
-                try {
-                    bfr.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-        return stringBuilder.toString();
-    }
-
-    private static String externalSDCardPath() {
-        String sdCardPath = "";
-        try {
-            StorageManager storageManager = (StorageManager) Coinlib.sInstance.getContext().getSystemService(Context.STORAGE_SERVICE);
-
-            // Android N started to have this method
-            List<StorageVolume> storageVolumes = storageManager.getStorageVolumes();
-            Class<?> volumeClass = Class.forName("android.os.storage.StorageVolume");
-            Method getPath = volumeClass.getDeclaredMethod("getPath");
-            getPath.setAccessible(true);
-            StorageVolume storageVolume = storageVolumes.get(storageVolumes.size() - 1);
-            sdCardPath = (String) getPath.invoke(storageVolume);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return sdCardPath;
     }
 
 }
